Streamlined the coding for level 1, Support doll ultimates in functions inside a parent class
